SSCS-7515 SSCS-7516 Merging in latest changes to refactor branch
