+++ src/main/java/uk/gov/hmcts/reform/sscs/service/DecisionNoticeOutcomeService.java
@@ -13,76 +13,21 @@ import uk.gov.hmcts.reform.sscs.ccd.presubmit.writefinaldecision.AwardType;
 import uk.gov.hmcts.reform.sscs.domain.wrapper.ComparedRate;
 
 @Slf4j
-@Service
-public class DecisionNoticeOutcomeService {
+public abstract class DecisionNoticeOutcomeService {
 
-    public Outcome determineOutcome(SscsCaseData sscsCaseData) {
+    private String benefitType;
 
-        if (sscsCaseData.getWriteFinalDecisionIsDescriptorFlow() == null) {
-            // We need at least this flag to be set in order to determine outcome
-            return null;
-        } else {
-            if (sscsCaseData.isDailyLivingAndOrMobilityDecision()) {
-                if ("yes".equalsIgnoreCase(sscsCaseData.getWriteFinalDecisionGenerateNotice())) {
-                    // If we are generating the notice we use the daily living/mobility descriptors
-                    // to determine outcome
-                    return determineGenerateNoticeDailyLivingOrMobilityFlowOutcome(sscsCaseData);
-                } else {
-                    // If we are not generating the notice we use an explicitly set outcome
-                    return useExplicitySetOutcome(sscsCaseData);
-                }
-            } else {
-                // If we are in the non-descriptor flow we use an explicitly set outcome.
-                return useExplicitySetOutcome(sscsCaseData);
-            }
-        }
+    protected DecisionNoticeOutcomeService(String benefitType) {
+        this.benefitType = benefitType;
     }
 
-    private Outcome determineGenerateNoticeDailyLivingOrMobilityFlowOutcome(SscsCaseData sscsCaseData) {
-
-        // Daily living and or/mobility
-
-        if ((!AwardType.NOT_CONSIDERED.getKey().equalsIgnoreCase(sscsCaseData.getPipWriteFinalDecisionDailyLivingQuestion())
-            && sscsCaseData.getPipWriteFinalDecisionComparedToDwpDailyLivingQuestion() == null)
-            || (!AwardType.NOT_CONSIDERED.getKey().equalsIgnoreCase(sscsCaseData.getPipWriteFinalDecisionMobilityQuestion())
-            && sscsCaseData.getPipWriteFinalDecisionComparedToDwpMobilityQuestion() == null)) {
-            return null;
-        } else {
-
-            try {
-
-                ComparedRate dailyLivingComparedRate = AwardType.NOT_CONSIDERED.getKey()
-                    .equalsIgnoreCase(sscsCaseData.getPipWriteFinalDecisionDailyLivingQuestion()) ? null :
-                    ComparedRate.getByKey(sscsCaseData.getPipWriteFinalDecisionComparedToDwpDailyLivingQuestion());
-
-                ComparedRate mobilityComparedRate = AwardType.NOT_CONSIDERED.getKey()
-                    .equalsIgnoreCase(sscsCaseData.getPipWriteFinalDecisionMobilityQuestion()) ? null : ComparedRate.getByKey(sscsCaseData.getPipWriteFinalDecisionComparedToDwpMobilityQuestion());
-
-                Set<ComparedRate> comparedRates = new HashSet<>();
-                if (dailyLivingComparedRate != null) {
-                    comparedRates.add(dailyLivingComparedRate);
-                }
-                if (mobilityComparedRate != null) {
-                    comparedRates.add(mobilityComparedRate);
-                }
-
-                // At least one higher,  and non lower, means the decision is in favour of appellant
-                if (comparedRates.contains(ComparedRate.Higher)) {
-                    return DECISION_IN_FAVOUR_OF_APPELLANT;
-                } else {
-                    // Otherwise, decision upheld
-                    return DECISION_UPHELD;
-                }
-
-            } catch (IllegalArgumentException e) {
-                log.error(e.getMessage());
-                return null;
-            }
-
-        }
+    public String getBenefitType() {
+        return benefitType;
     }
 
-    private Outcome useExplicitySetOutcome(SscsCaseData sscsCaseData) {
+    public abstract Outcome determineOutcome(SscsCaseData sscsCaseData);
+
+    protected Outcome useExplicitySetOutcome(SscsCaseData sscsCaseData) {
         if (sscsCaseData.getWriteFinalDecisionAllowedOrRefused() == null) {
             return null;
         } else {
